requirements.txt
torch>=2.3.0
torchaudio>=2.3.0
transformers>=4.41.2
datasets>=2.19.0
accelerate>=0.31.0
bitsandbytes>=0.43.0
deepspeed>=0.14.2
soundfile>=0.12.1
numpy>=1.26.0
pydantic>=2.6.0
fastapi>=0.111.0
uvicorn>=0.30.0
onnx>=1.16.0
onnxruntime>=1.18.0
pyyaml>=6.0.1
scipy>=1.11.4
encodec>=0.1.1
tqdm>=4.66.0
